Split Sudoku class into seperate files
